requirements.txt
Django==1.8.4
django-analytical==1.0.0
django-debug-toolbar==1.3.2
django-markdown==0.8.4
Markdown==2.6.2
sqlparse==0.1.16
wheel==0.24.0
